requirements.txt
boto3==1.20.23
pygame==2.1.2
python-dotenv==0.19.2
